requirements.txt
pyintelbras
fastapi[standard]
python-dotenv